+++ deploy/upgrade.sh
@@ -17,6 +17,7 @@ ensure_layout
 sync_tree
 seed_runtime_data_from_fixtures
 install_config_if_missing
+normalize_country_data_config_paths
 install_optional_config_if_missing "config/strings.toml" "strings.toml"
 install_optional_config_if_missing "config/pycluster.local.toml.example" "pycluster.local.toml.example"
 validate_or_refresh_strings_toml